fix: resolve clippy lints and broken test code surfaced by CI
The new `clippy --all-features -D warnings` gate flagged pre-existing
issues that previously went unnoticed because the policast-core test
module did not even compile:

- policy_store.rs: add the missing `principal_contract` field to three
  test `PolicyManifest` initializers (lib test failed to compile), and
  drop an unused `count()` helper.
- store.rs: remove `clone()` on `Copy` type `Effect`; fix doc list
  indentation.
- seed.rs: remove a redundant `i64 -> i64` cast.
- cel_filter.rs: collapse nested `if let` into a single pattern.

+++ crates/policast-uc/src/seed.rs
@@ -436,7 +436,7 @@ fn parse_ts_micros(s: &str) -> Option<i64> {
     let ss: u32 = ss_and_frac.split('.').next()?.parse().ok()?;
     // naive UTC -> micros via a tiny portable routine.
     let days = days_from_civil(y, m, d);
-    let secs = days as i64 * 86_400 + hh as i64 * 3_600 + mm as i64 * 60 + ss as i64;
+    let secs = days * 86_400 + hh as i64 * 3_600 + mm as i64 * 60 + ss as i64;
     Some(secs * 1_000_000)
 }
